# ent/http/delete.go
// Copyright 2019-present Facebook
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//
// Code generated by entc, DO NOT EDIT.

package http

import (
	"net/http"
	"strconv"

	"entgo.io/quynguyen-todo/ent"
	"github.com/gofiber/fiber/v2"
	"github.com/valyala/fasthttp/fasthttpadaptor"
	"go.uber.org/zap"
)

// @Summary Delete a Category
// @Description Delete a Category
// @ID Delete a Category
// @Tags [Category]
// @Accept  json
// @Produce  json
// @Param id path int true "Category ID"
// @Success 200 {object} ent.Category
// @Failure 400 {object} main.HTTPError
// @Failure 404 {object} main.HTTPError
// @Failure 500 {object} main.HTTPError
// @Router /category/{id} [delete]
// Delete removes a ent.Category from the database.
func (h CategoryHandler) Delete(c *fiber.Ctx) error {
	l := h.log.With(zap.String("method", "Delete"))
	// ID is URL parameter.
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		l.Error("error getting id from url parameter", zap.String("id", c.Params("id")), zap.Error(err))
		return c.Status(400).SendString("id must be an integer greater zero")
	}
	var r http.Request
	fasthttpadaptor.ConvertRequest(c.Context(), &r, true)
	err = h.client.Category.DeleteOneID(id).Exec(r.Context())
	if err != nil {
		switch {
		case ent.IsNotFound(err):
			msg := stripEntError(err)
			l.Info(msg, zap.Error(err), zap.Int("id", id))
			c.Status(404).SendString(msg)
		default:
			l.Error("could-not-delete-category", zap.Error(err), zap.Int("id", id))
			c.Status(fiber.StatusInternalServerError).SendString("Serve Error")
		}
		return nil
	}
	l.Info("category deleted", zap.Int("id", id))
	return c.SendString("Delete successfully")
}

// @Summary Delete a Product
// @Description Delete a Product
// @ID Delete a Product
// @Tags [Product]
// @Accept  json
// @Produce  json
// @Param id path int true "Product ID"
// @Success 200 {object} ent.Product
// @Failure 400 {object} main.HTTPError
// @Failure 404 {object} main.HTTPError
// @Failure 500 {object} main.HTTPError
// @Router /product/{id} [delete]
// Delete removes a ent.Product from the database.
func (h ProductHandler) Delete(c *fiber.Ctx) error {
	l := h.log.With(zap.String("method", "Delete"))
	// ID is URL parameter.
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		l.Error("error getting id from url parameter", zap.String("id", c.Params("id")), zap.Error(err))
		return c.Status(400).SendString("id must be an integer greater zero")
	}
	var r http.Request
	fasthttpadaptor.ConvertRequest(c.Context(), &r, true)
	err = h.client.Product.DeleteOneID(id).Exec(r.Context())
	if err != nil {
		switch {
		case ent.IsNotFound(err):
			msg := stripEntError(err)
			l.Info(msg, zap.Error(err), zap.Int("id", id))
			c.Status(404).SendString(msg)
		default:
			l.Error("could-not-delete-product", zap.Error(err), zap.Int("id", id))
			c.Status(fiber.StatusInternalServerError).SendString("Serve Error")
		}
		return nil
	}
	l.Info("product deleted", zap.Int("id", id))
	return c.SendString("Delete successfully")
}

// @Summary Delete a Todo
// @Description Delete a Todo
// @ID Delete a Todo
// @Tags [Todo]
// @Accept  json
// @Produce  json
// @Param id path int true "Todo ID"
// @Success 200 {object} ent.Todo
// @Failure 400 {object} main.HTTPError
// @Failure 404 {object} main.HTTPError
// @Failure 500 {object} main.HTTPError
// @Router /todo/{id} [delete]
// Delete removes a ent.Todo from the database.
func (h TodoHandler) Delete(c *fiber.Ctx) error {
	l := h.log.With(zap.String("method", "Delete"))
	// ID is URL parameter.
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		l.Error("error getting id from url parameter", zap.String("id", c.Params("id")), zap.Error(err))
		return c.Status(400).SendString("id must be an integer greater zero")
	}
	var r http.Request
	fasthttpadaptor.ConvertRequest(c.Context(), &r, true)
	err = h.client.Todo.DeleteOneID(id).Exec(r.Context())
	if err != nil {
		switch {
		case ent.IsNotFound(err):
			msg := stripEntError(err)
			l.Info(msg, zap.Error(err), zap.Int("id", id))
			c.Status(404).SendString(msg)
		default:
			l.Error("could-not-delete-todo", zap.Error(err), zap.Int("id", id))
			c.Status(fiber.StatusInternalServerError).SendString("Serve Error")
		}
		return nil
	}
	l.Info("todo deleted", zap.Int("id", id))
	return c.SendString("Delete successfully")
}

// @Summary Delete a VerySecret
// @Description Delete a VerySecret
// @ID Delete a VerySecret
// @Tags [VerySecret]
// @Accept  json
// @Produce  json
// @Param id path int true "VerySecret ID"
// @Success 200 {object} ent.VerySecret
// @Failure 400 {object} main.HTTPError
// @Failure 404 {object} main.HTTPError
// @Failure 500 {object} main.HTTPError
// @Router /verysecret/{id} [delete]
// Delete removes a ent.VerySecret from the database.
func (h VerySecretHandler) Delete(c *fiber.Ctx) error {
	l := h.log.With(zap.String("method", "Delete"))
	// ID is URL parameter.
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		l.Error("error getting id from url parameter", zap.String("id", c.Params("id")), zap.Error(err))
		return c.Status(400).SendString("id must be an integer greater zero")
	}
	var r http.Request
	fasthttpadaptor.ConvertRequest(c.Context(), &r, true)
	err = h.client.VerySecret.DeleteOneID(id).Exec(r.Context())
	if err != nil {
		switch {
		case ent.IsNotFound(err):
			msg := stripEntError(err)
			l.Info(msg, zap.Error(err), zap.Int("id", id))
			c.Status(404).SendString(msg)
		default:
			l.Error("could-not-delete-very-secret", zap.Error(err), zap.Int("id", id))
			c.Status(fiber.StatusInternalServerError).SendString("Serve Error")
		}
		return nil
	}
	l.Info("very-secret deleted", zap.Int("id", id))
	return c.SendString("Delete successfully")
}